(405) blocked user list and some refactoring
Created "blocked users" which has a list of all users currently blocked. Users have the option to unblock as well.

Added this page to the profile page, and redirected as so.

Refactored some code in profile page.

Added some padding in the CreateTrainingSessionForm, because the footer nav was hiding the bottom buttons.

Also changed the route to the Home page from "/HomePage" to "/" so made the necessary adjustment throughout the files.