feat(103): implement signup endpoints (#506)
* Renamed file authService to authAPI

To follow the naming convention of the other files under the Services/API folder. This has caused the authService.ts file to get deleted and it automatically created a new file authAPI with the same code. Also, updated import in useLogin.ts hook.

* Replaced HeaderNav with SecondaryHeader & removed FooterNav compoenent

* Add SecondaryHeader to SignUp page

* Added a postal code field

* Added a "success" variant

* Connecting the SignUp page to its backend using the API endpoints. Created 2 new files to implement hooks for the signup and verification code page.

* feat(103): created new hooks and implemented them in SignUp.tsx

created a hook for seconding the verification code. Modified the type to only accept Player, Coach or Admin.

* Formatted using Prettier

* Fixed error thrown in the lint check

* Fixed formatting with Prettier

* Added 2 slots to to accept a 6 pin code

* Format with prettier